backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]
pyjwt
pwdlib[argon2]